Make RoleException a proper class
